Add files via upload
Tic-Tac-Toe JavaScript game is a simple example of games you can program in JavaScript. Game making is the most popular branch of programming. It is also an amazing way to learn the basics of a programming language while creating a simple and interesting project. Updating cells for who wins the game of tic tac toe using arrays, java script, HTML, and CSS.